Move deployment scheme to schemes folder
